Added timeout for points. Moved mouse follow code
Added a flashing animation for when the points disappear. This is mainly
to speed up performance on slower computers, as it can slow down the
game to have too many particles flying around.

The code to shoot a bullet has been moved from the MouseFollow component
to the Player component, for sanity.